rename pgk to pkg
