+++ cmd/minectl/delete.go
@@ -1,7 +1,7 @@
 package minectl
 
 import (
-	"github.com/minectl/pgk/provisioner"
+	"github.com/minectl/pkg/provisioner"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
