+++ pkg/cloud/gce/gce.go
@@ -10,11 +10,11 @@ import (
 	"strings"
 	"time"
 
-	"github.com/minectl/pgk/update"
+	"github.com/minectl/pkg/update"
 
-	"github.com/minectl/pgk/automation"
-	"github.com/minectl/pgk/common"
-	minctlTemplate "github.com/minectl/pgk/template"
+	"github.com/minectl/pkg/automation"
+	"github.com/minectl/pkg/common"
+	minctlTemplate "github.com/minectl/pkg/template"
 	"github.com/pkg/errors"
 	"google.golang.org/api/compute/v1"
 	"google.golang.org/api/option"
