+++ pkg/cloud/hetzner/hetzner.go
@@ -9,12 +9,12 @@ import (
 	"strings"
 	"time"
 
-	"github.com/minectl/pgk/update"
+	"github.com/minectl/pkg/update"
 
 	"github.com/hetznercloud/hcloud-go/hcloud"
-	"github.com/minectl/pgk/automation"
-	"github.com/minectl/pgk/common"
-	minctlTemplate "github.com/minectl/pgk/template"
+	"github.com/minectl/pkg/automation"
+	"github.com/minectl/pkg/common"
+	minctlTemplate "github.com/minectl/pkg/template"
 )
 
 type Hetzner struct {
