+++ pkg/manifest/manifest.go
@@ -7,7 +7,7 @@ import (
 	"io/ioutil"
 	"log"
 
-	"github.com/minectl/pgk/model"
+	"github.com/minectl/pkg/model"
 	"github.com/xeipuuv/gojsonschema"
 	"sigs.k8s.io/yaml"
 )
