+++ pkg/manifest/schema.json
