+++ pkg/template/templates/bash/bash.tmpl
