+++ pkg/template/templates/bash/bedrock-binary.tmpl
