+++ pkg/template/templates/bash/fabric-binary.tmpl
