+++ pkg/template/templates/bash/forge-binary.tmpl
