+++ pkg/template/templates/bash/java-binary.tmpl
