+++ pkg/template/templates/bash/monitoring-binaries.tmpl
