+++ pkg/template/templates/bash/monitoring-files.tmpl
