+++ pkg/template/templates/bash/papermc-binary.tmpl
