+++ pkg/template/templates/bash/spigotbukkit-binary.tmpl
