+++ pkg/template/templates/cloud-init/bedrock-binary.yaml.tmpl
