+++ pkg/template/templates/cloud-init/cloud-config.yaml.tmpl
