+++ pkg/template/templates/cloud-init/fabric-binary.yaml.tmpl
