+++ pkg/template/templates/cloud-init/forge-binary.yaml.tmpl
