+++ pkg/template/templates/cloud-init/java-binary.yaml.tmpl
