+++ pkg/template/templates/cloud-init/monitoring-binaries.yaml.tmpl
