+++ pkg/template/templates/cloud-init/monitoring-files.yaml.tmpl
