+++ pkg/template/templates/cloud-init/papermc-binary.yaml.tmpl
