+++ pkg/template/templates/cloud-init/spigotbukkit-binary.yaml.tmpl
